commit-review/requirements.txt
openai==1.82.0
python-dotenv==1.1.0
